# Make sure no op updates don't affect the resource version due to serverside apply