**What type of PR is this?**
/kind bug
/priority important-soon
/sig api-machinery
/wg apply
/cc @apelisse

**What this PR does / why we need it**:
Adds a test to make sure no op updates don't affect the resource version due to serverside apply, and fixes a bug where this wasn't true

**Does this PR introduce a user-facing change?**:
```release-note
NONE
```