@jennybuckley 2019-08-20
@apelisse
```
jennybuckley@jennybuckley:~/Projects/k8s-p3/src/k8s.io/kubernetes$ benchcmp ~/old.bench ~/new.bench
benchmark                       old ns/op     new ns/op     delta
BenchmarkApplyNewObject-12      1613996       1517044       -6.01%
BenchmarkUpdateNewObject-12     1794913       2118134       +18.01%
BenchmarkRepeatedUpdate-12      636785        691073        +8.53%
BenchmarkSetToFields-12         5298          5212          -1.62%
BenchmarkFieldsToSet-12         10932         10716         -1.98%
BenchmarkYAMLToTyped-12         56229         57950         +3.06%

benchmark                       old allocs     new allocs     delta
BenchmarkApplyNewObject-12      6826           6826           +0.00%
BenchmarkUpdateNewObject-12     7893           10436          +32.22%
BenchmarkRepeatedUpdate-12      2246           2665           +18.66%
BenchmarkSetToFields-12         14             14             +0.00%
BenchmarkFieldsToSet-12         82             82             +0.00%
BenchmarkYAMLToTyped-12         152            152            +0.00%

benchmark                       old bytes     new bytes     delta
BenchmarkApplyNewObject-12      482317        482338        +0.00%
BenchmarkUpdateNewObject-12     534374        704804        +31.89%
BenchmarkRepeatedUpdate-12      189837        222702        +17.31%
BenchmarkSetToFields-12         1016          1016          +0.00%
BenchmarkFieldsToSet-12         6419          6419          +0.00%
BenchmarkYAMLToTyped-12         27405         27413         +0.03%
```