@jennybuckley 2019-08-20
I'm guessing mostly because compare is still allocating a merged object which we throw away