@jennybuckley 2019-08-20
And the other benchmarks:
```
$ benchcmp ~/old.bench ~/new.bench
benchmark                                                    old ns/op     new ns/op     delta
BenchmarkNoServerSideApply/List1-12                          1181222       1176036       -0.44%
BenchmarkNoServerSideApply/List20-12                         2284292       2251594       -1.43%
BenchmarkNoServerSideApply/List200-12                        12482119      12458413      -0.19%
BenchmarkNoServerSideApply/List2000-12                       80365197      80411328      +0.06%
BenchmarkNoServerSideApply/RepeatedUpdates-12                3398614       2946557       -13.30%
BenchmarkNoServerSideApply/Post1-12                          4098853       4022709       -1.86%
BenchmarkNoServerSideApply/Post10-12                         8309809       8184137       -1.51%
BenchmarkNoServerSideApply/Post50-12                         23449642      24398193      +4.05%
BenchmarkNoServerSideApplyButSameSize/List1-12               1242427       1208009       -2.77%
BenchmarkNoServerSideApplyButSameSize/List20-12              2854599       2799189       -1.94%
BenchmarkNoServerSideApplyButSameSize/List200-12             20112737      19981010      -0.65%
BenchmarkNoServerSideApplyButSameSize/List2000-12            132777713     152350769     +14.74%
BenchmarkNoServerSideApplyButSameSize/RepeatedUpdates-12     4203279       4203382       +0.00%
BenchmarkNoServerSideApplyButSameSize/Post1-12               4934311       5024805       +1.83%
BenchmarkNoServerSideApplyButSameSize/Post10-12              9418578       9921928       +5.34%
BenchmarkNoServerSideApplyButSameSize/Post50-12              27826752      27878165      +0.18%
BenchmarkServerSideApply/List1-12                            1340855       1413259       +5.40%
BenchmarkServerSideApply/List20-12                           4851633       4955054       +2.13%
BenchmarkServerSideApply/List200-12                          34224169      34746427      +1.53%
BenchmarkServerSideApply/List2000-12                         224618376     246000197     +9.52%
BenchmarkServerSideApply/RepeatedUpdates-12                  5522425       6037619       +9.33%
BenchmarkServerSideApply/Post1-12                            6540882       7112155       +8.73%
BenchmarkServerSideApply/Post10-12                           11559305      11591143      +0.28%
BenchmarkServerSideApply/Post50-12                           39453454      41130685      +4.25%

benchmark                                                    old allocs     new allocs     delta
BenchmarkNoServerSideApply/List1-12                          593            593            +0.00%
BenchmarkNoServerSideApply/List20-12                         4720           4720           +0.00%
BenchmarkNoServerSideApply/List200-12                        43679          43677          -0.00%
BenchmarkNoServerSideApply/List2000-12                       432775         432770         -0.00%
BenchmarkNoServerSideApply/RepeatedUpdates-12                1907           1907           +0.00%
BenchmarkNoServerSideApply/Post1-12                          5445           5445           +0.00%
BenchmarkNoServerSideApply/Post10-12                         54862          54860          -0.00%
BenchmarkNoServerSideApply/Post50-12                         274599         274480         -0.04%
BenchmarkNoServerSideApplyButSameSize/List1-12               608            608            +0.00%
BenchmarkNoServerSideApplyButSameSize/List20-12              4925           4926           +0.02%
BenchmarkNoServerSideApplyButSameSize/List200-12             45731          45734          +0.01%
BenchmarkNoServerSideApplyButSameSize/List2000-12            453372         453395         +0.01%
BenchmarkNoServerSideApplyButSameSize/RepeatedUpdates-12     1964           1968           +0.20%
BenchmarkNoServerSideApplyButSameSize/Post1-12               5572           5571           -0.02%
BenchmarkNoServerSideApplyButSameSize/Post10-12              56136          56132          -0.01%
BenchmarkNoServerSideApplyButSameSize/Post50-12              280779         280767         -0.00%
BenchmarkServerSideApply/List1-12                            630            630            +0.00%
BenchmarkServerSideApply/List20-12                           5367           5366           -0.02%
BenchmarkServerSideApply/List200-12                          50159          50164          +0.01%
BenchmarkServerSideApply/List2000-12                         497353         497459         +0.02%
BenchmarkServerSideApply/RepeatedUpdates-12                  8258           9496           +14.99%
BenchmarkServerSideApply/Post1-12                            10922          12268          +12.32%
BenchmarkServerSideApply/Post10-12                           109567         122975         +12.24%
BenchmarkServerSideApply/Post50-12                           548079         615141         +12.24%

benchmark                                                    old bytes     new bytes     delta
BenchmarkNoServerSideApply/List1-12                          52494         52428         -0.13%
BenchmarkNoServerSideApply/List20-12                         559171        559138        -0.01%
BenchmarkNoServerSideApply/List200-12                        5491381       5440290       -0.93%
BenchmarkNoServerSideApply/List2000-12                       51660430      51660935      +0.00%
BenchmarkNoServerSideApply/RepeatedUpdates-12                153417        153462        +0.03%
BenchmarkNoServerSideApply/Post1-12                          588174        587805        -0.06%
BenchmarkNoServerSideApply/Post10-12                         5995108       5995144       +0.00%
BenchmarkNoServerSideApply/Post50-12                         30036257      30004028      -0.11%
BenchmarkNoServerSideApplyButSameSize/List1-12               75205         75151         -0.07%
BenchmarkNoServerSideApplyButSameSize/List20-12              996043        995795        -0.02%
BenchmarkNoServerSideApplyButSameSize/List200-12             10615280      10640515      +0.24%
BenchmarkNoServerSideApplyButSameSize/List2000-12            101045079     101049658     +0.00%
BenchmarkNoServerSideApplyButSameSize/RepeatedUpdates-12     217272        218164        +0.41%
BenchmarkNoServerSideApplyButSameSize/Post1-12               680315        679998        -0.05%
BenchmarkNoServerSideApplyButSameSize/Post10-12              6901938       6900597       -0.02%
BenchmarkNoServerSideApplyButSameSize/Post50-12              34526224      34503709      -0.07%
BenchmarkServerSideApply/List1-12                            79463         79408         -0.07%
BenchmarkServerSideApply/List20-12                           1091447       1091469       +0.00%
BenchmarkServerSideApply/List200-12                          11554903      11658318      +0.89%
BenchmarkServerSideApply/List2000-12                         110666228     110665348     -0.00%
BenchmarkServerSideApply/RepeatedUpdates-12                  769688        871287        +13.20%
BenchmarkServerSideApply/Post1-12                            1117885       1234671       +10.45%
BenchmarkServerSideApply/Post10-12                           11278539      12415726      +10.08%
BenchmarkServerSideApply/Post50-12                           56475775      62137058      +10.02%
```