@jennybuckley 2019-08-20
Okay, without the comparison, the new benchcmp is much better:
```
$ benchcmp ~/old.bench ~/new.bench
benchmark                       old ns/op     new ns/op     delta
BenchmarkApplyNewObject-12      1625103       1569003       -3.45%
BenchmarkUpdateNewObject-12     1779586       1685061       -5.31%
BenchmarkRepeatedUpdate-12      628458        606244        -3.53%
BenchmarkSetToFields-12         5094          5156          +1.22%
BenchmarkFieldsToSet-12         10543         10078         -4.41%
BenchmarkYAMLToTyped-12         55939         55581         -0.64%

benchmark                       old allocs     new allocs     delta
BenchmarkApplyNewObject-12      6826           6826           +0.00%
BenchmarkUpdateNewObject-12     7893           7936           +0.54%
BenchmarkRepeatedUpdate-12      2246           2249           +0.13%
BenchmarkSetToFields-12         14             14             +0.00%
BenchmarkFieldsToSet-12         82             82             +0.00%
BenchmarkYAMLToTyped-12         152            152            +0.00%

benchmark                       old bytes     new bytes     delta
BenchmarkApplyNewObject-12      482346        482326        -0.00%
BenchmarkUpdateNewObject-12     534381        536068        +0.32%
BenchmarkRepeatedUpdate-12      189852        189906        +0.03%
BenchmarkSetToFields-12         1016          1016          +0.00%
BenchmarkFieldsToSet-12         6419          6419          +0.00%
BenchmarkYAMLToTyped-12         27409         27415         +0.02%
```